WindowManager changes after bindService flags extended to 64 bits.
Bug: 191785864
Test: atest cts/tests/app/src/android/app/cts/ServiceTest.java#testBindServiceFlags
atest cts/tests/app/src/android/app/cts/ServiceTest.java
atest cts/tests/app/src/android/app/cts/ActivityManagerProcessStateTest.java
atest cts/tests/app/src/android/app/cts/ActivityManagerFgsBgStartTest.java

+++ services/core/java/com/android/server/wm/BackgroundLaunchProcessController.java
@@ -241,7 +241,7 @@ class BackgroundLaunchProcessController {
         }
     }
 
-    void addBoundClientUid(int clientUid, String clientPackageName, int bindFlags) {
+    void addBoundClientUid(int clientUid, String clientPackageName, long bindFlags) {
         if (!CompatChanges.isChangeEnabled(
                 DEFAULT_RESCIND_BAL_FG_PRIVILEGES_BOUND_SERVICE,
                 clientPackageName,
